Issue 197 - detect and display ip information on /about page
